lib: hal: Add support for 64bit ARCH

+++ lib/hal/inc/hal_def.h
@@ -63,9 +63,9 @@
 
 #define HAL_IS_ALIGNED(x, a) (((x) & (a - 1)) == 0)
 #ifdef CACHE_LINE_SIZE
-#define HAL_IS_CACHELINE_ALIGNED(x) HAL_IS_ALIGNED((uint32_t)(x), CACHE_LINE_SIZE)
+#define HAL_IS_CACHELINE_ALIGNED(x) HAL_IS_ALIGNED((uintptr_t)(x), CACHE_LINE_SIZE)
 #else
-#define HAL_IS_CACHELINE_ALIGNED(x) HAL_IS_ALIGNED((uint32_t)(x), 4)
+#define HAL_IS_CACHELINE_ALIGNED(x) HAL_IS_ALIGNED((uintptr_t)(x), 4)
 #endif
 
 /* Compiller Macro */
